fast 下载功能，以及各种小功能

+++ app/src/main/java/com/a10miaomiao/bilimiao/entity/UserInfo.kt
@@ -26,7 +26,7 @@ data class UserInfo (
         var mid: Long,
         var name: String,
         var face: String,
-        var coin: Int,
+        var coin: Double,
         var bcoin: Double,
         var sex: Int,
         var rank: Int,
